Major improvements on dialogue area
* Add TextMeshProRuby to support Japanese kanji.
* By using TMP_Typewriter_With_Ruby, we can create typing text effect
easily and more control over them + 'With_Ruby' to support Japanese
kanji.
* The story will not be progress if the mouse pointer isn't on top of
Progress Story Button component. It's to make sure while we type
command in Debug Console and we pressed one of the Progress Story Keys
(Space, Return, etc.), the story will not be progress.
* Fixes the dialogue skip to the next dialogue when we trigger Progress
Story Keys or tap the screen and the dialogue is currently being typed
instead showing the rest of the dialogue.
* Replacing dialogue indicator animation with Tween.
* Changing dialogue typing interval to dialogue typing speed, it's also
effecting their command. From 'text interval' to 'text speed'.
The default dialogue typing speed is 25.
* Slightly changes in the UI.
* Add Japanese font. Currently only for testing.
* Update demo preview image.
* Update README.

Signed-off-by: Bayu Satiyo <itsyuukunz@gmail.com>